backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.46
psycopg2-binary==2.9.9

# API Clients & Data
httpx==0.25.2
aiohttp==3.9.1
stripe==7.0.0

# AI/LLM
anthropic==0.78.0
openai==1.26.0
google-generativeai==0.6.0

# ML/Data Science (For future phases)
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
ta==0.10.2

# Security & Authentication
python-jose==3.3.0
cryptography==42.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
